# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "smsService API",
        "description": "Send and receive SMS with smsService API",
        "version": "0.5.2",
        "termsOfService": "http://api.lineberty.net/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.lineberty.net/support",
            "email": "support@lineberty.net"
        },
        "license": {
            "name": "MIT License (MIT)",
            "url": "http://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8082",
    "schemes": [
        "http"
    ],
    "basePath": "/api/v1",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/sms": {
            "post": {
                "tags": [
                    "sms"
                ],
                "summary": "Send SMS to the phone number",
                "description": "",
                "operationId": "sendSms",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "The authorization token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "informations needed to send the sms",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sendSmsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid parameters supplied"
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                }
            }
        },
        "/callbacks/": {
            "post": {
                "tags": [
                    "callback"
                ],
                "summary": "Subscribe to get the receiver response",
                "description": "The callback (serviceUrl and callbackPath) is a POST route which accepts to receive an <a href='/#/definitions/incomingSms'>#incomingSms</a> object",
                "operationId": "subscribe",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "The authorization token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "informations needed to subscribe",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/subscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid parameters supplied"
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                }
            }
        },
        "/callbacks/phone/{phone}/serviceUrl/{serviceUrl}": {
            "delete": {
                "tags": [
                    "callback"
                ],
                "summary": "Unsubscribe to get a response from a phone",
                "description": "",
                "operationId": "unsubscribe",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "The authorization token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "phone",
                        "description": "The phone number",
                        "type": "string",
                        "pattern": "^00[0-9]{1,3}[0-9]{7,14}$",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "serviceUrl",
                        "description": "the service url of the callback",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid parameters supplied"
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                }
            }
        }
    },
    "definitions": {
        "sendSmsRequest": {
            "required": [
                "phone",
                "msg"
            ],
            "properties": {
                "phone": {
                    "type": "string",
                    "pattern": "^00[0-9]{1,3}[0-9]{7,14}$"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "subscribeRequest": {
            "required": [
                "phone",
                "serviceUrl",
                "callbackPath",
                "description"
            ],
            "properties": {
                "phone": {
                    "type": "string",
                    "pattern": "^00[0-9]{1,3}[0-9]{7,14}$"
                },
                "serviceUrl": {
                    "type": "string"
                },
                "callbackPath": {
                    "type": "string",
                    "description": "<a href='/#/definitions/incomingSms>incomingSms</a> is the object returned when a sms is received.'"
                },
                "description": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                }
            }
        },
        "incomingSms": {
            "required": [
                "phone",
                "msg",
                "time"
            ],
            "properties": {
                "phone": {
                    "type": "string",
                    "pattern": "^00[0-9]{1,3}[0-9]{7,14}$"
                },
                "msg": {
                    "type": "string",
                    "description": "message of incoming sms"
                },
                "description" : {
                    "type": "string",
                    "description": "description of the subscription defined with subscribe route"
                },
                "time" : {
                    "type": "number",
                    "description": "time when the sms was received by the provider"
                }
            }
        }
    }
}